Tried many different ways to get the scaling to work properly. Ended up by changed the zooming code to what is on master. Now the keys work fine and the scaling is good. - Need to apply this to the touch.
Also added an eagle icon for the user to select as a touch substitute of Q.

#story[1216]

+++ Visualizer/src/main/java/utilities/DataSource.java
@@ -18,7 +18,7 @@ import java.util.Set;
  */
 public interface DataSource {
     void changeScaling(double Multiplier);
-    double getZoomLevel();
+    int getZoomLevel();
     Map<Integer, CrewLocation> getCrewLocations();
     boolean receive(String host, int port, Scene scene);
     List<MutablePoint> getCourseBoundary();
@@ -36,7 +36,7 @@ public interface DataSource {
     Map<Integer, CourseFeature> getCourseFeatureMap();
     Map<Integer, List<Integer>> getIndexToSourceIdCourseFeatures();
     List<Double> getGPSbounds();
-    double getMapZoomLevel();
+    int getMapZoomLevel();
     double getShiftDistance();
     int getSourceID();
     Set<Integer> getCollisions();
